separate events and commands from main
