--- contributors.json
@@ -1,1 +1,0 @@
-{"contributors": [{"name": "Lapras", "uid": "395761182939807744"}, {"name": "Balu", "uid": "353572599957291010"}, {"name": "Pizzacat", "uid": "303732860265693185"}, {"name": "Spaghetto", "uid": "406302927884386318"}, {"name": "Baguette", "uid": "548974112131776522"}, {"name": "Breeze", "uid": "154033306894073856"}, {"name": "sarahtonein", "uid": "316765092689608706"}], "emojiIdMapping": {"<:sarah:1176399164154851368>": "316765092689608706", "<:lap:1110862059647795240>": "395761182939807744", "<:balu:1110862230343397387>": "353572599957291010", "<:pizzacat:1110862947145760939>": "303732860265693185", "<:spag:1110866508017586187>": "406302927884386318", "<:baguette:1113326206947954849>": "548974112131776522", "<:breeze:1113744843575922708>": "154033306894073856", "<:gunta:1176681877931839539>": "483209815212949515"}}
